requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pillow
pydub
sqlalchemy
python-multipart
edge-tts
